# src/swagger.json
{
    "openapi": "3.0.1",
    "info": { "title": "Users API", "version": "0.0.1" },
    "servers": [{ "url": "http://127.0.0.1:5000/api/v1" }],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": ["auth"],
                "summary": "Method for registering a new user",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "Register a user",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": ["auth"],
                "summary": "Method to login a user",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Login a user",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": ["auth"],
                "summary": "Method to logout a user",
                "operationId": "logoutUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": ["user"],
                "summary": "Method for fetching a list of users",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserOut"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }]
            },
            "post": {
                "tags": ["user"],
                "summary": "Method for creting a new user",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Register a user",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }]
            }
        },
        "/users/{userId}": {
            "get": {
                "tags": ["user"],
                "summary": "Find user by ID",
                "description": "Returns a single User",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of user to return",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserOut"
                                }
                            }
                        }
                    }
                },
                "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }]
            },
            "put": {
                "tags": ["user"],
                "summary": "Update user by ID",
                "description": "Updates a single User",
                "operationId": "updateUserById",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of user to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update a user",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserOut"
                                }
                            }
                        }
                    }
                },
                "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }]
            },
            "delete": {
                "tags": ["user"],
                "summary": "Delete a user by ID",
                "description": "Deletes a single User",
                "operationId": "deleteUserById",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of user to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserOut"
                                }
                            }
                        }
                    }
                },
                "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }]
            }
        }
    },
    "components": {
        "schemas": {
            "UserIn": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "john.doe@email.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password"
                    }
                },
                "xml": {
                    "name": "user"
                }
            },
            "UserUpdateIn": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "john.doe@email.com"
                    },
                    "active": {
                        "type": "boolean",
                        "example": "false"
                    }
                },
                "xml": {
                    "name": "user"
                }
            },
            "UserOut": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "uuid",
                        "example": "d0b0d0b0-d0b0-d0b0-d0b0-d0b0d0b0d0b0"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@email.com"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-01T00:00:00.000Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2020-01-01T00:00:00.000Z"
                    }
                },
                "xml": {
                    "name": "user"
                }
            },
            "ApiResponse": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "type": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "xml": {
                    "name": "##default"
                }
            }
        },
        "requestBodies": {
            "User": {
                "description": "User object",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UserIn"
                        }
                    }
                }
            },
            "Users": {
                "description": "List of user objects",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/components/schemas/UserIn"
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "jwt"
            }
        }
    }
}